fix: centralize multimodal tensor transport selection
