+++ test/registered/unit/multimodal/test_tensor_transport_mode.py
@@ -2,8 +2,9 @@
 
 import unittest
 from types import SimpleNamespace
+from unittest.mock import patch
 
-from sglang.srt.managers.mm_utils import determine_tensor_transport_mode
+from sglang.srt.multimodal.transport import determine_tensor_transport_mode
 from sglang.test.ci.ci_register import register_cpu_ci
 from sglang.test.test_utils import CustomTestCase
 
@@ -11,22 +12,27 @@ register_cpu_ci(est_time=1, suite="base-a-test-cpu")
 
 
 class TestTensorTransportMode(CustomTestCase):
-    def test_single_node_rendezvous_address_keeps_local_transport(self):
-        server_args = SimpleNamespace(
-            nnodes=1,
-            tp_size=2,
-            dist_init_addr="127.0.0.1:20000",
+    def test_transport_mode_uses_published_node_topology(self):
+        cases = (
+            (1, None, "cuda_ipc"),
+            (1, "127.0.0.1:20000", "cuda_ipc"),
+            (2, None, "default"),
+            (2, "10.0.0.1:20000", "default"),
         )
 
-        self.assertEqual(determine_tensor_transport_mode(server_args), "cuda_ipc")
-
-    def test_multi_node_uses_default_transport_without_address_heuristic(self):
-        server_args = SimpleNamespace(
-            nnodes=2,
-            dist_init_addr=None,
-        )
-
-        self.assertEqual(determine_tensor_transport_mode(server_args), "default")
+        for nnodes, dist_init_addr, expected in cases:
+            with self.subTest(nnodes=nnodes, dist_init_addr=dist_init_addr):
+                parallel = SimpleNamespace(
+                    nnodes=nnodes,
+                    dist_init_addr=dist_init_addr,
+                )
+                with patch(
+                    "sglang.srt.multimodal.transport.get_parallel",
+                    return_value=parallel,
+                ):
+                    self.assertEqual(
+                        determine_tensor_transport_mode(SimpleNamespace()), expected
+                    )
 
 
 if __name__ == "__main__":
